flask_server/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
SQLAlchemy==1.4.27
psycopg2-binary==2.9.2
Werkzeug==2.2.2